services/local_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
